add basic methods for virtual terminal

+++ utf8.cc
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <exception>
 
-unsigned short codepointSize(char byte) {
+int codepointSize(char byte) {
     if ((byte & 0x80) == 0) // 0xxxxxxx
         return 1;
     else if ((byte & 0xE0) == 0xC0) // 110xxxxx
@@ -12,6 +12,7 @@ unsigned short codepointSize(char byte) {
     else if ((byte & 0xF8) == 0xF0) // 11110xxx
         return 4;
 
+    // idk what to do if codepoint size is unknown
     return -1;
 }
